gf256: Move AVX2/SSSE3 code to separate sources
Although the gf256 implementation offers runtime detection of AVX2/SSSE3
capabilities, that alone does not suffice for portability. The main
problem is that, when a source code is compiled using the -mavx2 or
-mssse3 flags, the compiler is allowed to optimize other parts of the
code with AVX2/SSSE3 instruction sets, rather than using those
instructions only in the code section that intend to use them. As a
result, even though the runtime detection provided by CpuHasAVX2 and
CpuHasSSSE3 avoids calling unsupported code from the programmer's
perspective, there may still be other functions optimized for AVX2/SSSE3
that will ultimately lead to SIGILL.

To avoid the problem, this patch follows the strategy adopted for the
sha256 implementation. In particular, it moves the AVX2-specific code to
an isolated source (gf256_avx2.cpp) and compiles only that source with
the -mavx2 flag. Likewise, it moves the SSSE3-specific code to
gf256_ssse3.cpp and adds the -mssse3 flag only to that source. Then, the
base gf256.cpp implementation calls the specific bits from
gf256_avx2.cpp and gf256_ssse3.cpp as needed if those capabilities are
detected in runtime.